requirements.txt
tensorflow == 2.16.1
python-dotenv ~= 1.0.1
pillow ~= 10.2.0
matplotlib ~= 3.8.3
numpy == 1.26.4
keras~=3.1.1
opencv-python~=4.9.0.80